refactor: remove ui folder in components, put LoginForm in organisms
